Remove maximum dependency versions
It is common practice to use only minimum gem versions in gemspecs to
allow the gem to be "self maintaining" to some degree. We can apply a
maximum versions if a known incompatibility is found.

Allows rollie to be compatible with redis gem version 4 or future
versions of redis-namespace or connection_pool.